% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_LacCore_score.R
\docType{data}
\name{LACCORE_GENES}
\alias{LACCORE_GENES}
\title{The Canonical Six-Gene LacCore Signature}
\format{
A character vector of length 6.
}
\usage{
LACCORE_GENES
}
\description{
The Canonical Six-Gene LacCore Signature
}
\keyword{datasets}
