% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_LacCore_score.R
\docType{data}
\name{LRG_PANEL}
\alias{LRG_PANEL}
\title{The Full Panel of Lactylation-Related Genes (LRGs)}
\format{
A named list with five elements.
}
\usage{
LRG_PANEL
}
\description{
The Full Panel of Lactylation-Related Genes (LRGs)
}
\keyword{datasets}
